requirements.txt
python-dotenv==1.0.1
streamlit==1.41.1
